# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TelltaleDNS API",
    "description": "Read and control a TelltaleDNS resolver. JSON is camelCase; field names carry units (`totalMs`, `ttlSeconds`). Errors are RFC 9457 problem+json with a stable `code` and a `hint`. Time parameters take RFC 3339 or a relative offset (`-24h`). Collections page with `cursor`/`limit`. Analytics calls take `scope` (`cluster` or `node:local`; both mean this node until clustering is available).",
    "license": {
      "name": "Apache-2.0 OR MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/clients/{name}": {
      "put": {
        "tags": [
          "config"
        ],
        "summary": "Create, rename, or change a device (API-010).",
        "description": "Names devices across every view: the query log, live tail, charts, and history (names\nare resolved when read, so earlier queries are relabelled too). To rename, send the new\n`name` in the body. Devices defined in the config files are read-only here (409). Needs\nthe operator role (or a `write` token).",
        "operationId": "put_client",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "The device's current name (or the name for a new one).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dryRun",
            "in": "query",
            "description": "Validate and report the change without applying it.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Applied (or, with dryRun, what would change).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientChange"
                }
              }
            }
          },
          "409": {
            "description": "Defined in the config files, or an Idempotency-Key reused for another request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "412": {
            "description": "If-Match: the config changed since that version.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "The resulting configuration is invalid (e.g. an unknown group).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "config"
        ],
        "summary": "Delete a device named through the API.",
        "description": "Its queries, past and future, go back to showing the address, and it leaves its groups\n(the `default` group applies). Devices from the config files can't be deleted here (409).\nNeeds the operator role (or a `write` token).",
        "operationId": "delete_client",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "The device's name.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dryRun",
            "in": "query",
            "description": "Validate and report the change without applying it.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientChange"
                }
              }
            }
          },
          "404": {
            "description": "No device by that name was created through the API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Defined in the config files.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/queries": {
      "get": {
        "tags": [
          "queries"
        ],
        "summary": "Search the query log.",
        "description": "Newest first. Filters combine with AND: name (substring by default; `match=exact|suffix|\nglob|regex`), client, status, query type, response code, upstream, minimum latency, and\ntime range. Pass `nextCursor` back as `cursor` for older rows. Example:\n`GET /api/v1/queries?name=doubleclick&status=blocked&from=-1h&limit=50`.",
        "operationId": "queries",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Name to look for; see `match`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "match",
            "in": "query",
            "description": "How to match `name`. Default: `substring`.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/NameMatch"
            }
          },
          {
            "name": "client",
            "in": "query",
            "description": "Client IP address.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Statuses, comma-separated (`blocked,forwarded`).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "qtype",
            "in": "query",
            "description": "Query types, comma-separated (`A,AAAA`).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rcode",
            "in": "query",
            "description": "Response codes, comma-separated (`NXDOMAIN,SERVFAIL`).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "upstream",
            "in": "query",
            "description": "Upstream ID.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "minLatencyMs",
            "in": "query",
            "description": "Only queries that took at least this long.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Start: RFC 3339 or relative (`-1h`). Default: no limit.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End. Default: now.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Rows per page (1–1000, default 100).",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Continue after a previous page (`nextCursor` from it).",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "description": "`cluster` (default) or `node:local`.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryPage"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "The query log is off.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClientChange": {
        "type": "object",
        "description": "What a device change did, or would do with `dryRun=true` (AGT-002).",
        "required": [
          "applied",
          "configVersion",
          "recentQueries",
          "warnings"
        ],
        "properties": {
          "after": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ClientInfo",
                "description": "The device after the change (absent after a delete)."
              },
              {
                "type": "null"
              }
            ]
          },
          "applied": {
            "type": "boolean",
            "description": "False for a dry run."
          },
          "before": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ClientInfo",
                "description": "The device before the change (absent when it's new)."
              },
              {
                "type": "null"
              }
            ]
          },
          "configVersion": {
            "type": "integer",
            "format": "int64",
            "description": "The configuration version after the change (send it as `If-Match` next time).",
            "minimum": 0
          },
          "recentQueries": {
            "type": "integer",
            "format": "int64",
            "description": "Queries in the current and previous hour from the addresses it matches: how much\nhistory gets the new label (names are resolved when read, so the past is relabelled).",
            "minimum": 0
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Configuration warnings after the change."
          }
        }
      },
      "ClientInfo": {
        "type": "object",
        "description": "A configured client (device).",
        "required": [
          "name",
          "match",
          "groups",
          "source"
        ],
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Highest priority first."
          },
          "match": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "IPs, CIDRs, MACs, or `id:<client-id>` this device is recognized by."
          },
          "name": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "description": "`file` (defined in the config files: read-only here) or `api` (named in the UI or\nthrough the API: editable with `PUT /clients/{name}`).",
            "example": "api"
          }
        }
      },
      "ClientInput": {
        "type": "object",
        "description": "A device to create, rename, or change (`PUT /api/v1/clients/{name}`, API-010).",
        "required": [
          "match"
        ],
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Groups, highest priority first (default `[\"default\"]`)."
          },
          "match": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "How to recognize it: IPs, CIDRs, MACs (`aa:bb:cc:dd:ee:ff`), or `id:<client-id>`.",
            "example": [
              "192.168.1.42"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "The new name, to rename the device (default: the name in the path).",
            "example": "Living room TV"
          }
        },
        "additionalProperties": false
      },
      "Code": {
        "type": "string",
        "description": "Stable error codes.",
        "enum": [
          "invalid_parameter",
          "not_found",
          "unsupported_scope",
          "unavailable",
          "internal",
          "unauthorized",
          "totp_required",
          "forbidden",
          "csrf_rejected",
          "conflict",
          "version_conflict",
          "invalid_config",
          "rate_limited"
        ]
      },
...
      "NameMatch": {
        "type": "string",
        "description": "How `name` in [`QueryParams`] is matched.",
        "enum": [
          "substring",
          "exact",
          "suffix",
          "glob",
          "regex"
        ]
      },
...
      "Problem": {
        "type": "object",
        "description": "An API error.",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/Code",
            "description": "Stable machine-readable code."
          },
          "detail": {
            "type": "string",
            "description": "What went wrong in this request."
          },
          "hint": {
            "type": [
              "string",
              "null"
            ],
            "description": "What to do about it."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the problem type."
          },
          "type": {
            "type": "string",
            "description": "URI identifying the problem type."
          }
        },
        "example": {
          "code": "invalid_parameter",
          "detail": "`from`: `yesterday`: expected RFC 3339 time (2026-10-03T12:00:00Z) or a relative offset (-24h)",
          "hint": "Use a relative offset such as -24h, or RFC 3339 such as 2026-10-03T00:00:00Z.",
          "status": 400,
          "title": "Invalid parameter",
          "type": "https://telltaledns.dev/problems/invalid_parameter"
        }
      },
      "QueryPage": {
        "type": "object",
        "description": "A page of query-log rows, newest first.",
        "required": [
          "items",
          "scanned"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryRow"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Pass as `cursor` to get older rows; absent when there are none."
          },
          "scanned": {
            "$ref": "#/components/schemas/ScanStats",
            "description": "How much of the log this page looked at."
          }
        }
      },
      "QueryRow": {
        "type": "object",
        "description": "One logged query.",
        "required": [
          "time",
          "tsUnixMicros",
          "client",
          "name",
          "qtype",
          "status",
          "proto",
          "totalMs",
          "upstreamMs",
          "responseBytes",
          "answers"
        ],
        "properties": {
          "answers": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "client": {
            "type": "string",
            "description": "Client address."
          },
          "clientName": {
            "type": [
              "string",
              "null"
            ],
            "description": "The configured device name, when known."
          },
          "group": {
            "type": [
              "string",
              "null"
            ],
            "description": "The client's primary group."
          },
          "list": {
            "type": [
              "string",
              "null"
            ],
            "description": "The list whose rule blocked or allowed the query."
          },
          "name": {
            "type": "string"
          },
          "proto": {
            "type": "string"
          },
          "qtype": {
            "type": "string"
          },
          "rcode": {
            "type": [
              "string",
              "null"
            ],
            "description": "Response code, or null when nothing was sent."
          },
          "responseBytes": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "rule": {
            "type": [
              "string",
              "null"
            ],
            "description": "`domain`, `modifier`, `regex`, `cname`, or `allow`."
          },
          "status": {
            "type": "string"
          },
          "time": {
            "type": "string",
            "description": "When the query arrived (RFC 3339, milliseconds)."
          },
          "totalMs": {
            "type": "number",
            "format": "double"
          },
          "tsUnixMicros": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "upstreamMs": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "ScanStats": {
        "type": "object",
        "required": [
          "segments",
          "blocksRead",
          "blocksTotal",
          "rowsScanned"
        ],
        "properties": {
          "blocksRead": {
            "type": "integer",
            "minimum": 0
          },
          "blocksTotal": {
            "type": "integer",
            "minimum": 0
          },
          "rowsScanned": {
            "type": "integer",
            "minimum": 0
          },
          "segments": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "tt_<id>_<secret>"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "telltale_session"
      }
    }
  },
  "security": [
    {
      "session": []
    },
    {
      "bearer": []
    },
    {
      "basic": []
    }
  ],
...
}